refactor(memory_bridge): split tests.rs (#1266)
tests.rs 487 -> 243 LOC; new tests_extra.rs (~315 LOC).
error_bridge helper duplicated.
