+++ src/memory_bridge/mod.rs
@@ -386,3 +386,5 @@ struct ProspectivesResponse {
 
 #[cfg(test)]
 mod tests;
+#[cfg(test)]
+mod tests_extra;
